+++ api/urls.py
@@ -1,8 +1,11 @@
-from django.urls import path, include
-from . import views, kiosk_views
+from django.urls import include, path
+
+from . import kiosk_views, views
 
 kiosk_patterns = [
-    path("request_code/", kiosk_views.kiosk_request_code, name="api_kiosk_request_code"),
+    path(
+        "request_code/", kiosk_views.kiosk_request_code, name="api_kiosk_request_code"
+    ),
     path("unlock/", kiosk_views.kiosk_unlock, name="api_kiosk_unlock"),
     path("lock/", kiosk_views.kiosk_lock, name="api_kiosk_lock"),
     path("tap/", kiosk_views.kiosk_tap, name="api_kiosk_tap"),